+++ modules/apps/cookies/cookies-api/src/main/java/com/liferay/cookies/service/ConsentPreferenceServiceUtil.java
@@ -5,11 +5,7 @@
 
 package com.liferay.cookies.service;
 
-import com.liferay.cookies.model.ConsentPreference;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.module.service.Snapshot;
-
-import java.util.List;
 
 /**
  * Provides the remote service utility for ConsentPreference. This utility wraps
@@ -30,41 +26,6 @@ public class ConsentPreferenceServiceUtil {
 	 *
 	 * Never modify this class directly. Add custom service methods to <code>com.liferay.cookies.service.impl.ConsentPreferenceServiceImpl</code> and rerun ServiceBuilder to regenerate this class.
 	 */
-	public static ConsentPreference addConsentPreference(
-			long userId, String domain, java.util.Date expirationDate,
-			String name, String value)
-		throws PortalException {
-
-		return getService().addConsentPreference(
-			userId, domain, expirationDate, name, value);
-	}
-
-	public static void deleteConsentPreference(
-			long userId, String domain, String name)
-		throws PortalException {
-
-		getService().deleteConsentPreference(userId, domain, name);
-	}
-
-	public static void deleteConsentPreferences(long userId) {
-		getService().deleteConsentPreferences(userId);
-	}
-
-	public static void deleteConsentPreferences(long userId, String domain) {
-		getService().deleteConsentPreferences(userId, domain);
-	}
-
-	public static ConsentPreference getConsentPreference(
-		long userId, String domain, String name) {
-
-		return getService().getConsentPreference(userId, domain, name);
-	}
-
-	public static List<ConsentPreference> getConsentPreferences(
-		long userId, String domain) {
-
-		return getService().getConsentPreferences(userId, domain);
-	}
 
 	/**
 	 * Returns the OSGi service identifier.
@@ -73,12 +34,6 @@ public class ConsentPreferenceServiceUtil {
 	 */
 	public static String getOSGiServiceIdentifier() {
 		return getService().getOSGiServiceIdentifier();
-	}
-
-	public static ConsentPreference updateConsentPreference(
-		ConsentPreference consentPreference) {
-
-		return getService().updateConsentPreference(consentPreference);
 	}
 
 	public static ConsentPreferenceService getService() {
@@ -90,4 +45,4 @@ public class ConsentPreferenceServiceUtil {
 			ConsentPreferenceServiceUtil.class, ConsentPreferenceService.class);
 
 }
-// LIFERAY-SERVICE-BUILDER-HASH:2059967442
+// LIFERAY-SERVICE-BUILDER-HASH:253530810
