+++ modules/apps/cookies/cookies-service/src/main/java/com/liferay/cookies/service/http/ConsentPreferenceServiceHttp.java
@@ -5,21 +5,13 @@
 
 package com.liferay.cookies.service.http;
 
-import com.liferay.cookies.service.ConsentPreferenceServiceUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.security.auth.HttpPrincipal;
-import com.liferay.portal.kernel.service.http.TunnelUtil;
-import com.liferay.portal.kernel.util.MethodHandler;
-import com.liferay.portal.kernel.util.MethodKey;
-
 /**
  * Provides the HTTP utility for the
- * <code>ConsentPreferenceServiceUtil</code> service
+ * <code>com.liferay.cookies.service.ConsentPreferenceServiceUtil</code> service
  * utility. The
  * static methods of this class calls the same methods of the service utility.
  * However, the signatures are different because it requires an additional
- * <code>HttpPrincipal</code> parameter.
+ * <code>com.liferay.portal.kernel.security.auth.HttpPrincipal</code> parameter.
  *
  * <p>
  * The benefits of using the HTTP utility is that it is fast and allows for
@@ -40,263 +32,5 @@ import com.liferay.portal.kernel.util.MethodKey;
  * @generated
  */
 public class ConsentPreferenceServiceHttp {
-
-	public static com.liferay.cookies.model.ConsentPreference
-			addConsentPreference(
-				HttpPrincipal httpPrincipal, long userId, String domain,
-				java.util.Date expirationDate, String name, String value)
-		throws com.liferay.portal.kernel.exception.PortalException {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "addConsentPreference",
-				_addConsentPreferenceParameterTypes0);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, userId, domain, expirationDate, name, value);
-
-			Object returnObj = null;
-
-			try {
-				returnObj = TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				if (exception instanceof
-						com.liferay.portal.kernel.exception.PortalException) {
-
-					throw (com.liferay.portal.kernel.exception.PortalException)
-						exception;
-				}
-
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-
-			return (com.liferay.cookies.model.ConsentPreference)returnObj;
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static void deleteConsentPreference(
-			HttpPrincipal httpPrincipal, long userId, String domain,
-			String name)
-		throws com.liferay.portal.kernel.exception.PortalException {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "deleteConsentPreference",
-				_deleteConsentPreferenceParameterTypes1);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, userId, domain, name);
-
-			try {
-				TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				if (exception instanceof
-						com.liferay.portal.kernel.exception.PortalException) {
-
-					throw (com.liferay.portal.kernel.exception.PortalException)
-						exception;
-				}
-
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static void deleteConsentPreferences(
-		HttpPrincipal httpPrincipal, long userId) {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "deleteConsentPreferences",
-				_deleteConsentPreferencesParameterTypes2);
-
-			MethodHandler methodHandler = new MethodHandler(methodKey, userId);
-
-			try {
-				TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static void deleteConsentPreferences(
-		HttpPrincipal httpPrincipal, long userId, String domain) {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "deleteConsentPreferences",
-				_deleteConsentPreferencesParameterTypes3);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, userId, domain);
-
-			try {
-				TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static com.liferay.cookies.model.ConsentPreference
-		getConsentPreference(
-			HttpPrincipal httpPrincipal, long userId, String domain,
-			String name) {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "getConsentPreference",
-				_getConsentPreferenceParameterTypes4);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, userId, domain, name);
-
-			Object returnObj = null;
-
-			try {
-				returnObj = TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-
-			return (com.liferay.cookies.model.ConsentPreference)returnObj;
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static java.util.List<com.liferay.cookies.model.ConsentPreference>
-		getConsentPreferences(
-			HttpPrincipal httpPrincipal, long userId, String domain) {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "getConsentPreferences",
-				_getConsentPreferencesParameterTypes5);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, userId, domain);
-
-			Object returnObj = null;
-
-			try {
-				returnObj = TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-
-			return (java.util.List<com.liferay.cookies.model.ConsentPreference>)
-				returnObj;
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	public static com.liferay.cookies.model.ConsentPreference
-		updateConsentPreference(
-			HttpPrincipal httpPrincipal,
-			com.liferay.cookies.model.ConsentPreference consentPreference) {
-
-		try {
-			MethodKey methodKey = new MethodKey(
-				ConsentPreferenceServiceUtil.class, "updateConsentPreference",
-				_updateConsentPreferenceParameterTypes6);
-
-			MethodHandler methodHandler = new MethodHandler(
-				methodKey, consentPreference);
-
-			Object returnObj = null;
-
-			try {
-				returnObj = TunnelUtil.invoke(httpPrincipal, methodHandler);
-			}
-			catch (Exception exception) {
-				throw new com.liferay.portal.kernel.exception.SystemException(
-					exception);
-			}
-
-			return (com.liferay.cookies.model.ConsentPreference)returnObj;
-		}
-		catch (com.liferay.portal.kernel.exception.SystemException
-					systemException) {
-
-			_log.error(systemException, systemException);
-
-			throw systemException;
-		}
-	}
-
-	private static Log _log = LogFactoryUtil.getLog(
-		ConsentPreferenceServiceHttp.class);
-
-	private static final Class<?>[] _addConsentPreferenceParameterTypes0 =
-		new Class[] {
-			long.class, String.class, java.util.Date.class, String.class,
-			String.class
-		};
-	private static final Class<?>[] _deleteConsentPreferenceParameterTypes1 =
-		new Class[] {long.class, String.class, String.class};
-	private static final Class<?>[] _deleteConsentPreferencesParameterTypes2 =
-		new Class[] {long.class};
-	private static final Class<?>[] _deleteConsentPreferencesParameterTypes3 =
-		new Class[] {long.class, String.class};
-	private static final Class<?>[] _getConsentPreferenceParameterTypes4 =
-		new Class[] {long.class, String.class, String.class};
-	private static final Class<?>[] _getConsentPreferencesParameterTypes5 =
-		new Class[] {long.class, String.class};
-	private static final Class<?>[] _updateConsentPreferenceParameterTypes6 =
-		new Class[] {com.liferay.cookies.model.ConsentPreference.class};
-
 }
-// LIFERAY-SERVICE-BUILDER-HASH:-1328215813
+// LIFERAY-SERVICE-BUILDER-HASH:1026537262
