+++ modules/apps/cookies/cookies-service/src/main/java/com/liferay/cookies/service/persistence/impl/ConsentPreferencePersistenceImpl.java
@@ -361,159 +361,6 @@ public class ConsentPreferencePersistenceImpl
 	}
 
 	/**
-	 * Returns the consent preferences before and after the current consent preference in the ordered set where userId = &#63;.
-	 *
-	 * @param consentPreferenceId the primary key of the current consent preference
-	 * @param userId the user ID
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the previous, current, and next consent preference
-	 * @throws NoSuchConsentPreferenceException if a consent preference with the primary key could not be found
-	 */
-	@Override
-	public ConsentPreference[] findByUserId_PrevAndNext(
-			long consentPreferenceId, long userId,
-			OrderByComparator<ConsentPreference> orderByComparator)
-		throws NoSuchConsentPreferenceException {
-
-		ConsentPreference consentPreference = findByPrimaryKey(
-			consentPreferenceId);
-
-		Session session = null;
-
-		try {
-			session = openSession();
-
-			ConsentPreference[] array = new ConsentPreferenceImpl[3];
-
-			array[0] = getByUserId_PrevAndNext(
-				session, consentPreference, userId, orderByComparator, true);
-
-			array[1] = consentPreference;
-
-			array[2] = getByUserId_PrevAndNext(
-				session, consentPreference, userId, orderByComparator, false);
-
-			return array;
-		}
-		catch (Exception exception) {
-			throw processException(exception);
-		}
-		finally {
-			closeSession(session);
-		}
-	}
-
-	protected ConsentPreference getByUserId_PrevAndNext(
-		Session session, ConsentPreference consentPreference, long userId,
-		OrderByComparator<ConsentPreference> orderByComparator,
-		boolean previous) {
-
-		StringBundler sb = null;
-
-		if (orderByComparator != null) {
-			sb = new StringBundler(
-				4 + (orderByComparator.getOrderByConditionFields().length * 3) +
-					(orderByComparator.getOrderByFields().length * 3));
-		}
-		else {
-			sb = new StringBundler(3);
-		}
-
-		sb.append(_SQL_SELECT_CONSENTPREFERENCE_WHERE);
-
-		sb.append(_FINDER_COLUMN_USERID_USERID_2);
-
-		if (orderByComparator != null) {
-			String[] orderByConditionFields =
-				orderByComparator.getOrderByConditionFields();
-
-			if (orderByConditionFields.length > 0) {
-				sb.append(WHERE_AND);
-			}
-
-			for (int i = 0; i < orderByConditionFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByConditionFields[i]);
-
-				if ((i + 1) < orderByConditionFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN_HAS_NEXT);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN);
-					}
-				}
-			}
-
-			sb.append(ORDER_BY_CLAUSE);
-
-			String[] orderByFields = orderByComparator.getOrderByFields();
-
-			for (int i = 0; i < orderByFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByFields[i]);
-
-				if ((i + 1) < orderByFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC_HAS_NEXT);
-					}
-					else {
-						sb.append(ORDER_BY_DESC_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC);
-					}
-					else {
-						sb.append(ORDER_BY_DESC);
-					}
-				}
-			}
-		}
-		else {
-			sb.append(ConsentPreferenceModelImpl.ORDER_BY_JPQL);
-		}
-
-		String sql = sb.toString();
-
-		Query query = session.createQuery(sql);
-
-		query.setFirstResult(0);
-		query.setMaxResults(2);
-
-		QueryPos queryPos = QueryPos.getInstance(query);
-
-		queryPos.add(userId);
-
-		if (orderByComparator != null) {
-			for (Object orderByConditionValue :
-					orderByComparator.getOrderByConditionValues(
-						consentPreference)) {
-
-				queryPos.add(orderByConditionValue);
-			}
-		}
-
-		List<ConsentPreference> list = query.list();
-
-		if (list.size() == 2) {
-			return list.get(1);
-		}
-		else {
-			return null;
-		}
-	}
-
-	/**
 	 * Removes all the consent preferences where userId = &#63; from the database.
 	 *
 	 * @param userId the user ID
@@ -876,173 +723,6 @@ public class ConsentPreferencePersistenceImpl
 		}
 
 		return null;
-	}
-
-	/**
-	 * Returns the consent preferences before and after the current consent preference in the ordered set where expirationDate = &#63;.
-	 *
-	 * @param consentPreferenceId the primary key of the current consent preference
-	 * @param expirationDate the expiration date
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the previous, current, and next consent preference
-	 * @throws NoSuchConsentPreferenceException if a consent preference with the primary key could not be found
-	 */
-	@Override
-	public ConsentPreference[] findByExpirationDate_PrevAndNext(
-			long consentPreferenceId, Date expirationDate,
-			OrderByComparator<ConsentPreference> orderByComparator)
-		throws NoSuchConsentPreferenceException {
-
-		ConsentPreference consentPreference = findByPrimaryKey(
-			consentPreferenceId);
-
-		Session session = null;
-
-		try {
-			session = openSession();
-
-			ConsentPreference[] array = new ConsentPreferenceImpl[3];
-
-			array[0] = getByExpirationDate_PrevAndNext(
-				session, consentPreference, expirationDate, orderByComparator,
-				true);
-
-			array[1] = consentPreference;
-
-			array[2] = getByExpirationDate_PrevAndNext(
-				session, consentPreference, expirationDate, orderByComparator,
-				false);
-
-			return array;
-		}
-		catch (Exception exception) {
-			throw processException(exception);
-		}
-		finally {
-			closeSession(session);
-		}
-	}
-
-	protected ConsentPreference getByExpirationDate_PrevAndNext(
-		Session session, ConsentPreference consentPreference,
-		Date expirationDate,
-		OrderByComparator<ConsentPreference> orderByComparator,
-		boolean previous) {
-
-		StringBundler sb = null;
-
-		if (orderByComparator != null) {
-			sb = new StringBundler(
-				4 + (orderByComparator.getOrderByConditionFields().length * 3) +
-					(orderByComparator.getOrderByFields().length * 3));
-		}
-		else {
-			sb = new StringBundler(3);
-		}
-
-		sb.append(_SQL_SELECT_CONSENTPREFERENCE_WHERE);
-
-		boolean bindExpirationDate = false;
-
-		if (expirationDate == null) {
-			sb.append(_FINDER_COLUMN_EXPIRATIONDATE_EXPIRATIONDATE_1);
-		}
-		else {
-			bindExpirationDate = true;
-
-			sb.append(_FINDER_COLUMN_EXPIRATIONDATE_EXPIRATIONDATE_2);
-		}
-
-		if (orderByComparator != null) {
-			String[] orderByConditionFields =
-				orderByComparator.getOrderByConditionFields();
-
-			if (orderByConditionFields.length > 0) {
-				sb.append(WHERE_AND);
-			}
-
-			for (int i = 0; i < orderByConditionFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByConditionFields[i]);
-
-				if ((i + 1) < orderByConditionFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN_HAS_NEXT);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN);
-					}
-				}
-			}
-
-			sb.append(ORDER_BY_CLAUSE);
-
-			String[] orderByFields = orderByComparator.getOrderByFields();
-
-			for (int i = 0; i < orderByFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByFields[i]);
-
-				if ((i + 1) < orderByFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC_HAS_NEXT);
-					}
-					else {
-						sb.append(ORDER_BY_DESC_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC);
-					}
-					else {
-						sb.append(ORDER_BY_DESC);
-					}
-				}
-			}
-		}
-		else {
-			sb.append(ConsentPreferenceModelImpl.ORDER_BY_JPQL);
-		}
-
-		String sql = sb.toString();
-
-		Query query = session.createQuery(sql);
-
-		query.setFirstResult(0);
-		query.setMaxResults(2);
-
-		QueryPos queryPos = QueryPos.getInstance(query);
-
-		if (bindExpirationDate) {
-			queryPos.add(new Timestamp(expirationDate.getTime()));
-		}
-
-		if (orderByComparator != null) {
-			for (Object orderByConditionValue :
-					orderByComparator.getOrderByConditionValues(
-						consentPreference)) {
-
-				queryPos.add(orderByConditionValue);
-			}
-		}
-
-		List<ConsentPreference> list = query.list();
-
-		if (list.size() == 2) {
-			return list.get(1);
-		}
-		else {
-			return null;
-		}
 	}
 
 	/**
@@ -1441,179 +1121,6 @@ public class ConsentPreferencePersistenceImpl
 		}
 
 		return null;
-	}
-
-	/**
-	 * Returns the consent preferences before and after the current consent preference in the ordered set where userId = &#63; and domain = &#63;.
-	 *
-	 * @param consentPreferenceId the primary key of the current consent preference
-	 * @param userId the user ID
-	 * @param domain the domain
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the previous, current, and next consent preference
-	 * @throws NoSuchConsentPreferenceException if a consent preference with the primary key could not be found
-	 */
-	@Override
-	public ConsentPreference[] findByU_D_PrevAndNext(
-			long consentPreferenceId, long userId, String domain,
-			OrderByComparator<ConsentPreference> orderByComparator)
-		throws NoSuchConsentPreferenceException {
-
-		domain = Objects.toString(domain, "");
-
-		ConsentPreference consentPreference = findByPrimaryKey(
-			consentPreferenceId);
-
-		Session session = null;
-
-		try {
-			session = openSession();
-
-			ConsentPreference[] array = new ConsentPreferenceImpl[3];
-
-			array[0] = getByU_D_PrevAndNext(
-				session, consentPreference, userId, domain, orderByComparator,
-				true);
-
-			array[1] = consentPreference;
-
-			array[2] = getByU_D_PrevAndNext(
-				session, consentPreference, userId, domain, orderByComparator,
-				false);
-
-			return array;
-		}
-		catch (Exception exception) {
-			throw processException(exception);
-		}
-		finally {
-			closeSession(session);
-		}
-	}
-
-	protected ConsentPreference getByU_D_PrevAndNext(
-		Session session, ConsentPreference consentPreference, long userId,
-		String domain, OrderByComparator<ConsentPreference> orderByComparator,
-		boolean previous) {
-
-		StringBundler sb = null;
-
-		if (orderByComparator != null) {
-			sb = new StringBundler(
-				5 + (orderByComparator.getOrderByConditionFields().length * 3) +
-					(orderByComparator.getOrderByFields().length * 3));
-		}
-		else {
-			sb = new StringBundler(4);
-		}
-
-		sb.append(_SQL_SELECT_CONSENTPREFERENCE_WHERE);
-
-		sb.append(_FINDER_COLUMN_U_D_USERID_2);
-
-		boolean bindDomain = false;
-
-		if (domain.isEmpty()) {
-			sb.append(_FINDER_COLUMN_U_D_DOMAIN_3);
-		}
-		else {
-			bindDomain = true;
-
-			sb.append(_FINDER_COLUMN_U_D_DOMAIN_2);
-		}
-
-		if (orderByComparator != null) {
-			String[] orderByConditionFields =
-				orderByComparator.getOrderByConditionFields();
-
-			if (orderByConditionFields.length > 0) {
-				sb.append(WHERE_AND);
-			}
-
-			for (int i = 0; i < orderByConditionFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByConditionFields[i]);
-
-				if ((i + 1) < orderByConditionFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN_HAS_NEXT);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(WHERE_GREATER_THAN);
-					}
-					else {
-						sb.append(WHERE_LESSER_THAN);
-					}
-				}
-			}
-
-			sb.append(ORDER_BY_CLAUSE);
-
-			String[] orderByFields = orderByComparator.getOrderByFields();
-
-			for (int i = 0; i < orderByFields.length; i++) {
-				sb.append(_ORDER_BY_ENTITY_ALIAS);
-				sb.append(orderByFields[i]);
-
-				if ((i + 1) < orderByFields.length) {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC_HAS_NEXT);
-					}
-					else {
-						sb.append(ORDER_BY_DESC_HAS_NEXT);
-					}
-				}
-				else {
-					if (orderByComparator.isAscending() ^ previous) {
-						sb.append(ORDER_BY_ASC);
-					}
-					else {
-						sb.append(ORDER_BY_DESC);
-					}
-				}
-			}
-		}
-		else {
-			sb.append(ConsentPreferenceModelImpl.ORDER_BY_JPQL);
-		}
-
-		String sql = sb.toString();
-
-		Query query = session.createQuery(sql);
-
-		query.setFirstResult(0);
-		query.setMaxResults(2);
-
-		QueryPos queryPos = QueryPos.getInstance(query);
-
-		queryPos.add(userId);
-
-		if (bindDomain) {
-			queryPos.add(domain);
-		}
-
-		if (orderByComparator != null) {
-			for (Object orderByConditionValue :
-					orderByComparator.getOrderByConditionValues(
-						consentPreference)) {
-
-				queryPos.add(orderByConditionValue);
-			}
-		}
-
-		List<ConsentPreference> list = query.list();
-
-		if (list.size() == 2) {
-			return list.get(1);
-		}
-		else {
-			return null;
-		}
 	}
 
 	/**
@@ -2635,4 +2142,4 @@ public class ConsentPreferencePersistenceImpl
 	}
 
 }
-// LIFERAY-SERVICE-BUILDER-HASH:-1590664540
+// LIFERAY-SERVICE-BUILDER-HASH:1552982916
